product-review-analysis/requirements.txt
requests==2.31.0
numpy==1.24.3
pandas==2.0.1
matplotlib==3.7.1
seaborn==0.12.2
beautifulsoup4==4.12.3
python-dotenv==1.0.0